+++ color/src/androidTest/java/com/vanpra/composematerialdialogs/color/test/functional/ColorPickerDialogTests.kt
@@ -4,8 +4,8 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.semantics.SemanticsActions
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.performClick
-import androidx.compose.ui.test.performGesture
 import androidx.compose.ui.test.performSemanticsAction
+import androidx.compose.ui.test.performTouchInput
 import androidx.compose.ui.test.swipeLeft
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.vanpra.composematerialdialogs.color.ARGBPickerState
@@ -157,7 +157,7 @@ class ColorPickerDialogTests {
             }
         }
 
-        composeTestRule.onDialogColorPicker().performGesture { swipeLeft() }
+        composeTestRule.onDialogColorPicker().performTouchInput { swipeLeft() }
         composeTestRule.waitForIdle()
 
         composeTestRule