+++ core/src/main/java/com/vanpra/composematerialdialogs/MaterialDialogLists.kt
@@ -28,8 +28,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
 
-private val bottomPadding = Modifier.padding(bottom = 8.dp)
-
 /**
  * Adds a selectable list with custom items to the dialog
  *
@@ -40,14 +38,19 @@ private val bottomPadding = Modifier.padding(bottom = 8.dp)
  */
 @Composable
 fun <T> MaterialDialogScope.listItems(
+    modifier: Modifier = Modifier,
     list: List<T>,
     closeOnClick: Boolean = true,
     onClick: (index: Int, item: T) -> Unit = { _, _ -> },
     isEnabled: (index: Int) -> Boolean = { _ -> true },
-    item: @Composable (index: Int, T) -> Unit
+    item: @Composable (index: Int, T) -> Unit,
 ) {
     BoxWithConstraints {
-        LazyColumn(Modifier.then(bottomPadding).testTag("dialog_list")) {
+        LazyColumn(
+            Modifier
+                .then(modifier)
+                .testTag("dialog_list")
+        ) {
             itemsIndexed(list) { index, it ->
                 Box(
                     Modifier
@@ -62,7 +65,6 @@ fun <T> MaterialDialogScope.listItems(
                             },
                             enabled = isEnabled(index)
                         )
-                        .padding(horizontal = 24.dp)
                 ) {
                     item(index, it)
                 }
@@ -83,14 +85,19 @@ fun MaterialDialogScope.listItems(
     closeOnClick: Boolean = true,
     onClick: (index: Int, item: String) -> Unit = { _, _ -> }
 ) {
-    listItems(list = list, closeOnClick = closeOnClick, onClick = onClick) { _, item ->
+    listItems(
+        modifier = Modifier.padding(bottom = 8.dp),
+        list = list,
+        closeOnClick = closeOnClick,
+        onClick = onClick
+    ) { _, item ->
         Text(
             item,
             color = MaterialTheme.colors.onSurface,
             style = MaterialTheme.typography.body1,
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(vertical = 12.dp)
+                .padding(vertical = 12.dp, horizontal = 24.dp)
                 .wrapContentWidth(Alignment.Start)
         )
     }
@@ -224,7 +231,8 @@ private fun MultiChoiceItem(
     Row(
         Modifier
             .fillMaxWidth()
-            .height(48.dp),
+            .height(48.dp)
+            .padding(start = 12.dp, end = 24.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
         Checkbox(checked = selected, onCheckedChange = { onChecked(index) }, enabled = enabled)
@@ -256,7 +264,8 @@ private fun SingleChoiceItem(
     Row(
         Modifier
             .fillMaxWidth()
-            .height(48.dp),
+            .height(48.dp)
+            .padding(start = 12.dp, end = 24.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
         RadioButton(