+++ test-utils/src/main/java/com/vanpra/composematerialdialogs/test/utils/extensions/CoreComposeTestRule.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.test.SemanticsNodeInteraction
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.performGesture
+import androidx.compose.ui.test.performTouchInput
 import androidx.compose.ui.test.swipeUp
 
 fun ComposeContentTestRule.setContentAndWaitForIdle(content: @Composable () -> Unit) {
@@ -37,7 +37,7 @@ fun ComposeTestRule.onDialogListItem(index: Int): SemanticsNodeInteraction {
     try {
         onNodeWithTag("dialog_list_item_$index").assertExists()
     } catch (e: AssertionError) {
-        onDialogList().performGesture { swipeUp() }
+        onDialogList().performTouchInput { swipeUp() }
         waitForIdle()
     }